refactor(approximators): move Approximator trait to core
