+++ src/approximators/mod.rs
@@ -1,43 +1,6 @@
 //! Module for efficient dot product abstractions used by `LFA`.
-use crate::core::*;
-use crate::geometry::Matrix;
+use crate::{core::*, geometry::Matrix};
 use std::{collections::HashMap, mem::replace};
-
-/// An interface for function approximators.
-pub trait Approximator<I: ?Sized> {
-    type Value;
-
-    /// Return the dimensionality of the output value `Approximator::Value`.
-    fn n_outputs(&self) -> usize;
-
-    /// Evaluate the function and return its value.
-    fn evaluate(&self, input: &I) -> EvaluationResult<Self::Value>;
-
-    /// Update the approximator's estimate for the given input.
-    fn update(&mut self, input: &I, update: Self::Value) -> UpdateResult<()>;
-
-    #[allow(unused_variables)]
-    /// Adapt the approximator in light of newly discovered features.
-    fn adapt(&mut self, new_features: &HashMap<IndexT, IndexSet>) -> AdaptResult<usize> {
-        unimplemented!()
-    }
-}
-
-impl<I: ?Sized, T: Approximator<I>> Approximator<I> for Box<T> {
-    type Value = T::Value;
-
-    fn n_outputs(&self) -> usize { (**self).n_outputs() }
-
-    fn evaluate(&self, input: &I) -> EvaluationResult<Self::Value> { (**self).evaluate(input) }
-
-    fn update(&mut self, input: &I, update: Self::Value) -> UpdateResult<()> {
-        (**self).update(input, update)
-    }
-
-    fn adapt(&mut self, new_features: &HashMap<IndexT, IndexSet>) -> AdaptResult<usize> {
-        (**self).adapt(new_features)
-    }
-}
 
 fn append_matrix_rows(weights: &mut Matrix<f64>, new_rows: Vec<Vec<f64>>) {
     let n_cols = weights.cols();
