+++ src/core/mod.rs
@@ -1,4 +1,5 @@
 //! Module for common types and primitives used throughout the crate.
 import_all!(primitives);
 import_all!(error);
+import_all!(approximator);
 import_all!(parameterised);
